// src/app/core/models/permission.model.ts
export type Permission =
  | 'BRANCH_VIEW'
  | 'BRANCH_CREATE'
  | 'BRANCH_UPDATE'
  | 'BRANCH_DELETE'
  | 'TECHNICIAN_VIEW'
  | 'TECHNICIAN_CREATE'
  | 'TECHNICIAN_UPDATE'
  | 'TECHNICIAN_DELETE'
  | 'TECHNICIAN_ASSIGN'
  | 'INVENTORY_VIEW'
  | 'INVENTORY_CREATE'
  | 'INVENTORY_UPDATE'
  | 'INVENTORY_DELETE'
  | 'INVENTORY_RESERVE'
  | 'INVENTORY_CONSUME'
  | 'STOCK_MOVEMENT_VIEW'
  | 'STOCK_MOVEMENT_CREATE'
  | 'SERVICE_REQUEST_VIEW'
  | 'SERVICE_REQUEST_CREATE'
  | 'SERVICE_REQUEST_UPDATE'
  | 'SERVICE_REQUEST_DELETE'
  | 'WORK_ORDER_VIEW'
  | 'WORK_ORDER_CREATE'
  | 'WORK_ORDER_PLAN'
  | 'WORK_ORDER_UPDATE_STATUS'
  | 'WORK_ORDER_CANCEL'
  | 'WORK_ORDER_COMPLETE'
  | 'WORK_ORDER_PARTIAL_COMPLETE'
  | 'VEHICLE_VIEW'
  | 'VEHICLE_CREATE'
  | 'VEHICLE_UPDATE'
  | 'VEHICLE_DELETE'
  | 'RULE_VIEW'
  | 'RULE_MANAGE'
  | 'NOTIFICATION_VIEW'
  | 'NOTIFICATION_MANAGE'
  | 'REPORT_VIEW'
  | 'REPORT_EXPORT'
  | 'AUDIT_LOG_VIEW'
  | 'IMPORT_EXECUTE'
  | 'EXPORT_EXECUTE'
  | 'SIMULATION_RUN'
  | 'SETTINGS_MANAGE'
  | 'SHIFT_VIEW'
  | 'SHIFT_CREATE'
  | 'SHIFT_ASSIGN'
  | 'SHIFT_DELETE';